Move implementation of Pallets spec to a public package
