+++ internal/app/forklift/constraints.go
@@ -1,23 +1,6 @@
 package forklift
 
-// Conflicts
-
-func CheckResourcesConflicts[Resource ConflictChecker[Resource]](
-	first []AttachedResource[Resource], second []AttachedResource[Resource],
-) (conflicts []ResourceConflict[Resource]) {
-	for _, f := range first {
-		for _, s := range second {
-			if errs := f.Resource.CheckConflict(s.Resource); errs != nil {
-				conflicts = append(conflicts, ResourceConflict[Resource]{
-					First:  f,
-					Second: s,
-					Errs:   errs,
-				})
-			}
-		}
-	}
-	return conflicts
-}
+// DeplConflict
 
 func (c DeplConflict) HasNameConflict() bool {
 	return c.Name
@@ -40,7 +23,7 @@ func (c DeplConflict) HasConflict() bool {
 		c.HasListenerConflict() || c.HasNetworkConflict() || c.HasServiceConflict()
 }
 
-// Dependencies
+// MissingDeplDependencies
 
 func (c MissingDeplDependencies) HasMissingNetworkDependency() bool {
 	return len(c.Networks) > 0
@@ -52,33 +35,4 @@ func (c MissingDeplDependencies) HasMissingServiceDependency() bool {
 
 func (c MissingDeplDependencies) HasMissingDependency() bool {
 	return c.HasMissingNetworkDependency() || c.HasMissingServiceDependency()
-}
-
-func CheckResourcesDependencies[Resource DependencyChecker[Resource]](
-	required []AttachedResource[Resource], provided []AttachedResource[Resource],
-) (missingDeps []MissingResourceDependency[Resource]) {
-	for _, r := range required {
-		bestErrsCount := -1
-		bestCandidates := make([]ResourceDependencyCandidate[Resource], 0, len(provided))
-		for _, p := range provided {
-			errs := r.Resource.CheckDependency(p.Resource)
-			if bestErrsCount == -1 || len(errs) <= bestErrsCount {
-				if len(errs) < bestErrsCount {
-					bestCandidates = make([]ResourceDependencyCandidate[Resource], 0, len(provided))
-				}
-				bestErrsCount = len(errs)
-				bestCandidates = append(bestCandidates, ResourceDependencyCandidate[Resource]{
-					Provided: p,
-					Errs:     errs,
-				})
-			}
-		}
-		if bestErrsCount != 0 {
-			missingDeps = append(missingDeps, MissingResourceDependency[Resource]{
-				Required:       r,
-				BestCandidates: bestCandidates,
-			})
-		}
-	}
-	return missingDeps
 }
