+++ internal/app/forklift/models-constraints.go
@@ -1,21 +1,8 @@
 package forklift
 
-type AttachedResource[Resource interface{}] struct {
-	Resource Resource
-	Source   []string
-}
-
-// Conflict constraints
-
-type ConflictChecker[Resource interface{}] interface {
-	CheckConflict(candidate Resource) []error
-}
-
-type ResourceConflict[Resource interface{}] struct {
-	First  AttachedResource[Resource]
-	Second AttachedResource[Resource]
-	Errs   []error
-}
+import (
+	"github.com/PlanktoScope/forklift/pkg/pallets"
+)
 
 type DeplConflict struct {
 	First  Depl
@@ -23,30 +10,14 @@ type DeplConflict struct {
 
 	// Possible conflicts
 	Name      bool
-	Listeners []ResourceConflict[ListenerResource]
-	Networks  []ResourceConflict[NetworkResource]
-	Services  []ResourceConflict[ServiceResource]
-}
-
-// Dependency constraints
-
-type DependencyChecker[Resource interface{}] interface {
-	CheckDependency(candidate Resource) []error
-}
-
-type ResourceDependencyCandidate[Resource interface{}] struct {
-	Provided AttachedResource[Resource]
-	Errs     []error
-}
-
-type MissingResourceDependency[Resource interface{}] struct {
-	Required       AttachedResource[Resource]
-	BestCandidates []ResourceDependencyCandidate[Resource]
+	Listeners []pallets.ResourceConflict[pallets.ListenerResource]
+	Networks  []pallets.ResourceConflict[pallets.NetworkResource]
+	Services  []pallets.ResourceConflict[pallets.ServiceResource]
 }
 
 type MissingDeplDependencies struct {
 	Depl Depl
 
-	Networks []MissingResourceDependency[NetworkResource]
-	Services []MissingResourceDependency[ServiceResource]
+	Networks []pallets.MissingResourceDependency[pallets.NetworkResource]
+	Services []pallets.MissingResourceDependency[pallets.ServiceResource]
 }
